+++ tasty-plutus/CHANGELOG.md
@@ -12,6 +12,7 @@ This format is based on [Keep A Changelog](https://keepachangelog.com/en/1.0.0).
 * Multiple new modules.
 * Additional `tasty` options to control property-based execution.
 * More examples in documentation.
+* Dependency on `plutus-pretty`.
 
 ### Changed